DAY_3/enterprise_asset_mgmt/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
alembic
pytest
httpx
